.github/workflows/scripts/requirements.txt
pyYaml==6.0